pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "locallens"
version = "0.1.0"
description = "Search your files by talking to them — 100% offline semantic search with voice"
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.11,<3.14"
authors = [{name = "Mahimai"}]
keywords = ["semantic-search", "qdrant", "voice-assistant", "offline", "rag"]

dependencies = [
    "typer>=0.12.0",
    "rich>=13.0.0",
    "qdrant-client>=1.13.0",
    "qdrant-edge-py>=0.6.0",
    "sentence-transformers>=3.0.0",
    "pymupdf>=1.24.0",
    "python-docx>=1.1.0",
    "httpx>=0.27.0",
    "numpy>=1.26.0",
    "rank-bm25>=0.2.2",
]

[project.optional-dependencies]
voice = [
    "moonshine-voice>=0.0.50",
    "piper-tts>=1.4.0",
    "sounddevice>=0.5.0",
    "onnxruntime>=1.19.0",
]
parsing = [
    "liteparse>=0.1.0",
    "openpyxl>=3.1.0",
]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]
watch = [
    "watchdog>=4.0.0",
]
email = [
    "python-oletools>=0.60.0",
    "extract-msg>=0.48.0",
]
ebooks = [
    "ebooklib>=0.18",
]
test = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.0.0",
    "httpx>=0.27.0",
]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "anyio[trio]>=4.0.0",
    "httpx>=0.27.0",
    "ruff>=0.4.0",
]

[project.scripts]
locallens = "locallens.cli:app"

[project.entry-points."locallens.extractors"]
builtin_text = "locallens.extractors.text:TextExtractor"
builtin_pdf = "locallens.extractors.pdf:PdfExtractor"
builtin_docx = "locallens.extractors.docx_ext:DocxExtractor"
builtin_code = "locallens.extractors.code:CodeExtractor"
builtin_spreadsheet = "locallens.extractors.spreadsheet:SpreadsheetExtractor"
builtin_liteparse = "locallens.extractors.liteparse_ext:LiteParseExtractor"
builtin_email = "locallens.extractors.email_ext:EmailExtractor"
builtin_epub = "locallens.extractors.epub_ext:EpubExtractor"
builtin_obsidian = "locallens.extractors.obsidian_ext:ObsidianExtractor"

[tool.setuptools.packages.find]
include = ["locallens*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
markers = [
    "slow: marks tests that require Ollama or are long-running (deselect with '-m \"not slow\"')",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = [
    "E",
    "W",
    "F",
    "I",
    "B",
    "C4",
    "UP",
]
ignore = [
    "E501",   # line length — handled by formatter
    "B008",   # function call in default arg (FastAPI Depends pattern)
    "B904",   # raise from in except — CLI uses raise Exit() intentionally
    "B905",   # zip without strict — too pedantic
    "C901",   # function complexity
    "W191",   # tab indentation
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["F401", "F811"]
"locallens/voice.py" = ["B023"]

[tool.mypy]
python_version = "3.11"
warn_return_any = false
warn_unused_configs = true
ignore_missing_imports = true
explicit_package_bases = true

[[tool.mypy.overrides]]
module = [
    "locallens.voice",
    "locallens.extractors.*",
    "backend.app.services.watcher",
    "backend.app.services.voice_stt",
]
ignore_errors = true

[tool.pyright]
typeCheckingMode = "off"
